Create README.md
The objectives for this project is to create a semi-to-fully-functioning YouTube homepage clone, that includes video previews, a working drop-down menu,  a functioning search-bar, a functioning search button, and a functioning voice-search button( hopefully) among other things. Note: This is my first project working with HTML5 and CSS, therefore I am learning as I go.